chore: rename `*.config.json` to `*.datasource.config.json`
