+++ docs/site/tutorials/todo/todo-tutorial-geocoding-service.md
@@ -47,7 +47,7 @@ $ lb4 datasource
 ? Default options for the request:
 ? An array of operation templates:
 ? Use default CRUD mapping: No
-   create src/datasources/geocoder.config.json
+   create src/datasources/geocoder.datasource.config.json
    create src/datasources/geocoder.datasource.ts
  # npm will install dependencies now
     update src/datasources/index.ts
@@ -59,7 +59,7 @@ Edit the newly created datasource configuration to configure Geocoder API
 endpoints. Configuration options provided by REST Connector are described in our
 docs here: [REST connector](/doc/en/lb3/REST-connector.html).
 
-{% include code-caption.html content="/src/datasources/geocoder.config.json" %}
+{% include code-caption.html content="/src/datasources/geocoder.datasource.config.json" %}
 
 ```json
 {
