+++ examples/soap-calculator/src/datasources/calculator.datasource.config.json
