+++ examples/todo/src/__tests__/helpers.ts
@@ -6,7 +6,7 @@
 import {HttpCachingProxy} from '@loopback/http-caching-proxy';
 import {merge} from 'lodash';
 import * as path from 'path';
-import * as GEO_CODER_CONFIG from '../datasources/geocoder.config.json';
+import * as GEO_CODER_CONFIG from '../datasources/geocoder.datasource.config.json';
 import {Todo} from '../models/index';
 import {GeocoderService, GeoPoint} from '../services/geocoder.service';
