+++ examples/todo/src/datasources/db.datasource.ts
@@ -5,7 +5,7 @@
 
 import {inject} from '@loopback/core';
 import {juggler} from '@loopback/repository';
-import * as config from './db.config.json';
+import * as config from './db.datasource.config.json';
 
 export class DbDataSource extends juggler.DataSource {
   static dataSourceName = 'db';
