+++ examples/todo/src/datasources/geocoder.datasource.ts
@@ -5,7 +5,7 @@
 
 import {inject} from '@loopback/core';
 import {AnyObject, juggler} from '@loopback/repository';
-import * as config from './geocoder.config.json';
+import * as config from './geocoder.datasource.config.json';
 
 export class GeocoderDataSource extends juggler.DataSource {
   static dataSourceName = 'geocoder';
