+++ packages/cli/test/integration/generators/datasource.integration.js
@@ -64,7 +64,7 @@ const expectedTSFile = path.join(
 
 const expectedJSONFile = path.join(
   SANDBOX_PATH,
-  'src/datasources/ds.config.json',
+  'src/datasources/ds.datasource.config.json',
 );
 
 const expectedIndexFile = path.join(SANDBOX_PATH, 'src/datasources/index.ts');
